Added pipeline master file from prod-pipeline. prod-master creates provision lambda and new pipeline-master
The prod-master template no longer creates a prod-pipeline stack, but instead a pipeline-master stack.
This stack creates supporting resources needed for a pipeline, but does not actually create a pipeline itself.
Pipeline-master creates a cloudwatch event rule that matches all write events in source bucket, and triggers
the provisioning lambda.